backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
google-api-python-client==2.108.0
selenium==4.15.2
webdriver-manager==4.0.1
apscheduler==3.10.4
beautifulsoup4==4.12.2
pydantic==2.5.0
alembic==1.12.1
requests==2.31.0